Skip OPENSSL_cpuid_setup when replaying.
For RUN-1614 (https://linear.app/replay/issue/RUN-1614/skip-checking-cpuid-during-replay-in-boringssl-bypassing-openssl-cpuid)
