+++ crypto/CMakeLists.txt
@@ -337,6 +337,7 @@ add_library(
   rand_extra/rand_extra.c
   rand_extra/windows.c
   rc4/rc4.c
+  record_replay.c
   refcount_c11.c
   refcount_lock.c
   rsa_extra/rsa_asn1.c
