+++ crypto/record_replay.h
@@ -1,0 +1,7 @@
+#ifndef OPENSSL_HEADER_RECORD_REPLAY_H
+#define OPENSSL_HEADER_RECORD_REPLAY_H
+
+int RecordReplay_IsRecordingOrReplaying(void);
+int RecordReplay_IsReplaying(void);
+
+#endif // OPENSSL_HEADER_RECORD_REPLAY_H
